requirements.txt
streamlit
google-generativeai
python-dotenv
PyMuPDF
pdf2image
frontend
openai==0.28
pypdf2
